GF.Support

Supporting infrastructure and system utilities

Source locations

class HasSourcePath a where

Methods

sourcePath :: a -> FilePath

data Location

Constructors

NoLoc

Local Int Int

External FilePath Location

data L a

Attaching location information

Constructors

L Location a

unLoc :: L a -> a

noLoc :: a -> L a

ppLocation :: FilePath -> Location -> Doc

ppL :: (Pretty a2, Pretty a1) => L a2 -> a1 -> Doc

Command line options

Option types

data Options

data Flags

Constructors

Flags

Fields

optMode :: Mode

optStopAfterPhase :: Phase

optVerbosity :: Verbosity

optShowCPUTime :: Bool

optOutputFormats :: [OutputFormat]

optSISR :: Maybe SISRFormat

optHaskellOptions :: Set HaskellOption

optLexicalCats :: Set String

optLiteralCats :: Set Ident

optGFODir :: Maybe FilePath

optOutputDir :: Maybe FilePath

optGFLibPath :: Maybe FilePath

optDocumentRoot :: Maybe FilePath

optRecomp :: Recomp

optProbsFile :: Maybe FilePath

optRetainResource :: Bool

optName :: Maybe String

optPreprocessors :: [String]

optEncoding :: Maybe String

optPMCFG :: Bool

optOptimizations :: Set Optimization

optOptimizePGF :: Bool

optSplitPGF :: Bool

optCFGTransforms :: Set CFGTransform

optLibraryPath :: [FilePath]

optStartCat :: Maybe String

optSpeechLanguage :: Maybe String

optLexer :: Maybe String

optUnlexer :: Maybe String

optWarnings :: [Warning]

optDump :: [Dump]

optTagsOnly :: Bool

optHeuristicFactor :: Maybe Double

optPlusAsBind :: Bool

optJobs :: Maybe (Maybe Int)

data Mode

Constructors

ModeVersion

ModeHelp

ModeInteractive

ModeRun

ModeCompiler

ModeServer Int

data Phase

Constructors

Preproc

Convert

Compile

Link

data Verbosity

Constructors

Quiet

Normal

Verbose

Debug

data OutputFormat

Constructors

FmtPGFPretty

FmtJavaScript

FmtPython

FmtHaskell

FmtProlog

FmtLambdaProlog

FmtByteCode

FmtBNF

FmtEBNF

FmtRegular

FmtNoLR

FmtSRGS_XML

FmtSRGS_XML_NonRec

FmtSRGS_ABNF

FmtSRGS_ABNF_NonRec

FmtJSGF

FmtGSL

FmtVoiceXML

FmtSLF

FmtRegExp

FmtFA

data SISRFormat

Constructors

SISR_WD20030401
SISR Working draft 1 April 2003 http://www.w3.org/TR/2003/WD-semantic-interpretation-20030401/

SISR_1_0

data Optimization

Constructors

OptStem

OptCSE

OptExpand

OptParametrize

data CFGTransform

Constructors

CFGNoLR

CFGRegular

CFGTopDownFilter

CFGBottomUpFilter

CFGStartCatOnly

CFGMergeIdentical

CFGRemoveCycles

data HaskellOption

Constructors

HaskellNoPrefix

HaskellGADT

HaskellLexical

HaskellConcrete

HaskellVariants

newtype Dump

Constructors

Dump Pass

data Pass

Constructors

Source

Rebuild

Extend

Rename

TypeCheck

Refresh

Optimize

Canon

data Recomp

Constructors

AlwaysRecomp

RecompIfNewer

NeverRecomp

outputFormatsExpl :: [((String, OutputFormat), String)]

Option parsing

parseOptions

Arguments

:: ErrorMonad err

=> [String]
list of string arguments

-> err (Options, [FilePath])

parseModuleOptions

Arguments

:: ErrorMonad err

=> [String]
list of string arguments

-> err Options

fixRelativeLibPaths :: FilePath -> FilePath -> Options -> Options

Option pretty-printing

optionsGFO :: Options -> [(String, Literal)]

Pretty-print the options that are preserved in .gfo files.

optionsPGF :: Options -> [(String, Literal)]

Pretty-print the options that are preserved in .pgf files.

Option manipulation

addOptions :: Options -> Options -> Options

concatOptions :: [Options] -> Options

noOptions :: Options

modifyFlags :: (Flags -> Flags) -> Options

helpMessage :: String

Checking specific options

flag :: (Flags -> a) -> Options -> a

cfgTransform :: Options -> CFGTransform -> Bool

haskellOption :: Options -> HaskellOption -> Bool

readOutputFormat :: Monad m => String -> m OutputFormat

isLexicalCat :: Options -> String -> Bool

isLiteralCat :: Options -> Ident -> Bool

renameEncoding :: String -> String

This is for bacward compatibility. Since GHC 6.12 we started using the native Unicode support in GHC but it uses different names for the code pages.

getEncoding :: Options -> String

defaultEncoding :: [Char]

Setting specific options

setOptimization :: Optimization -> Bool -> Options

setCFGTransform :: CFGTransform -> Bool -> Options

Convenience methods for checking options

verbAtLeast :: Options -> Verbosity -> Bool

dump :: Options -> Dump -> Bool

The Error monad

data Err a

Like Maybe type with error msgs

Constructors

Ok a

Bad String

err :: (String -> b) -> (a -> b) -> Err a -> b

Analogue of maybe

maybeErr :: ErrorMonad m => String -> Maybe a -> m a

Add msg s to Maybe failures

testErr :: ErrorMonad m => Bool -> String -> m ()

fromErr :: a -> Err a -> a

Analogue of fromMaybe

errIn :: ErrorMonad m => String -> m a -> m a

lookupErr :: (ErrorMonad m, Eq a, Show a) => a -> [(a, b)] -> m b

Error monad class

class (Functor m, Monad m) => ErrorMonad m where

Minimal complete definition

raise, handle

Methods

raise :: String -> m a

handle :: m a -> (String -> m a) -> m a

handle_ :: m a -> m a -> m a

checks :: ErrorMonad m => [m a] -> m a

doUntil :: ErrorMonad m => (a -> Bool) -> [m a] -> m a

liftErr :: ErrorMonad m => Err a -> m a

Checking

checkUnique :: (Show a, Eq a) => [a] -> [String]

unifyMaybeBy :: (Eq b, Monad m) => (a -> b) -> Maybe a -> Maybe a -> m (Maybe a)

unifyMaybe :: (Eq a, Monad m) => Maybe a -> Maybe a -> m (Maybe a)

this is what happens when matching two values in the same module

Monadic operations on lists and pairs

mapPairListM :: Monad m => ((a, b) -> m c) -> [(a, b)] -> m [(a, c)]

mapPairsM :: Monad m => (b -> m c) -> [(a, b)] -> m [(a, c)]

pairM :: Monad m => (b -> m c) -> (b, b) -> m (c, c)

Binary search trees; now with FiniteMap

type BinTree a b = Map a b

emptyBinTree :: BinTree a b

isInBinTree :: Ord a => a -> BinTree a b -> Bool

lookupTree :: (ErrorMonad m, Ord a) => (a -> String) -> a -> BinTree a b -> m b

lookupTreeManyAll :: Ord a => (a -> String) -> [BinTree a b] -> a -> [b]

updateTree :: Ord a => (a, b) -> BinTree a b -> BinTree a b

buildTree :: Ord a => [(a, b)] -> BinTree a b

filterBinTree :: Ord a => (a -> b -> Bool) -> BinTree a b -> BinTree a b

mapTree :: ((a, b) -> c) -> BinTree a b -> BinTree a c

tree2list :: BinTree a b -> [(a, b)]

Printing

indent :: Int -> String -> String

(+++) :: String -> String -> String infixr 5

(++-) :: String -> String -> String infixr 5

(++++) :: String -> String -> String infixr 5

(+++++) :: String -> String -> String infixr 5

prUpper :: String -> String

prReplicate :: Int -> String -> String

prTList :: String -> [String] -> String

prQuotedString :: String -> String

prParenth :: String -> String

prCurly :: String -> String

prBracket :: String -> String

prArgList :: [String] -> String

prSemicList :: [String] -> String

prCurlyList :: [String] -> String

restoreEscapes :: String -> String

numberedParagraphs :: [[String]] -> [String]

prConjList :: String -> [String] -> String

prIfEmpty :: String -> String -> String -> String -> String

wrapLines :: Int -> String -> String

Thomas Hallgren's wrap lines

Topological sorting

topoTest :: Ord a => [(a, [a])] -> Either [a] [[a]]

Topological sorting with test of cyclicity

topoTest2 :: Ord a => [(a, [a])] -> Either [[a]] [[a]]

Topological sorting with test of cyclicity, new version /TH 2012-06-26

Misc

ifNull :: b -> ([a] -> b) -> [a] -> b

combinations :: [[a]] -> [[a]]

combinations is the same as sequence!!! peb 30/5-04

done :: Monad m => m ()

return ()

readIntArg :: String -> Int

iterFix :: Eq a => ([a] -> [a]) -> [a] -> [a]

Fix point iterator (for computing e.g. transitive closures or reachability)

chunks :: Eq a => a -> [a] -> [[a]]

chop into separator-separated parts

Files and IO

putIfVerb :: Output m => Options -> String -> m ()

GF files path and library path manipulation

type FileName = String

type InitPath

Arguments

= String
the directory portion of a pathname

type FullPath = String

gfLibraryPath :: [Char]

gfGrammarPathVar :: [Char]

getLibraryDirectory :: MonadIO io => Options -> io FilePath

getGrammarPath :: MonadIO io => FilePath -> io [FilePath]

extendPathEnv :: MonadIO io => Options -> io [FilePath]

extends the search path with the gfLibraryPath and gfGrammarPathVar environment variables. Returns only existing paths.

getSubdirs :: FilePath -> IO [FilePath]

justModuleName :: FilePath -> String

isGF :: FilePath -> Bool

isGFO :: FilePath -> Bool

gfFile :: FilePath -> FilePath

gfoFile :: FilePath -> FilePath

gf2gfo :: Options -> FilePath -> FilePath

gf2gfo' :: Maybe FilePath -> FilePath -> FilePath

splitInModuleSearchPath :: String -> [FilePath]

Error handling in the IO monad

type IOE a = IO a

Was: newtype IOE a = IOE { appIOE :: IO (Err a) }

tryIOE :: IOE a -> IO (Err a)

Catch exceptions caused by calls to raise or fail in the IO monad. To catch all IO exceptions, use try instead.

useIOE :: a -> IOE a -> IO a

Print the error message and return a default value if the IO operation fails

maybeIO :: (MonadIO f, Functor f) => IO a -> f (Maybe a)

die :: String -> IO a

Diagnostic output

class Monad m => Output m where

Methods

ePutStr, putStrLnE, putStrE, ePutStrLn :: String -> m ()

putPointE :: (Output m, MonadIO m) => Verbosity -> Options -> String -> m b -> m b

ioErrorText :: IOError -> String

Because GHC adds the confusing text "user error" for failures caused by calls to fail.

Timing

timeIt :: MonadIO m => m t -> m (Integer, t)

File IO

writeUTF8File :: FilePath -> String -> IO ()

readBinaryFile :: FilePath -> IO String

writeBinaryFile :: FilePath -> String -> IO ()

Reused

class Monad m => MonadIO m where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

liftErr :: ErrorMonad m => Err a -> m a

Backwards compatible try and catch

catch :: IO a -> (IOError -> IO a) -> IO a

try :: IO a -> IO (Either IOError a)

Console IO

Changing which character encoding to use for console IO

setConsoleEncoding :: IO ()

Set the console encoding (for Windows, has no effect on Unix-like systems)

changeConsoleEncoding :: String -> IO ()

Console colors

data TermColors

Constructors

TermColors

Fields

redFg :: String

blueFg :: String

restore :: String

getTermColors :: MonadIO m => m TermColors

Binary serialisation

class Binary t

The Binary class provides put and get, methods to encode and decode a Haskell value to a lazy ByteString. It mirrors the Read and Show classes for textual representation of Haskell types, and is suitable for serialising Haskell values to disk, over the network.

For parsing and generating simple external binary formats (e.g. C structures), Binary may be used, but in general is not suitable for complex protocols. Instead use the Put and Get primitives directly.

Instances of Binary should satisfy the following property:

decode . encode == id

That is, the get and put methods should be the inverse of each other. A range of instances are provided for basic Haskell types.

Minimal complete definition

put, get

encode :: Binary a => a -> ByteString

Encode a value using binary serialisation to a lazy ByteString.

decode :: Binary a => ByteString -> a

Decode a value from a lazy ByteString, reconstructing the original structure.

encodeFile :: Binary a => FilePath -> a -> IO ()

Lazily serialise a value to a file

This is just a convenience function, it's defined simply as:

encodeFile f = B.writeFile f . encode

So for example if you wanted to compress as well, you could use:

B.writeFile f . compress . encode

decodeFile :: Binary a => FilePath -> IO a

Lazily reconstruct a value previously written to a file.

This is just a convenience function, it's defined simply as:

decodeFile f = return . decode =<< B.readFile f

So for example if you wanted to decompress as well, you could use:

return . decode . decompress =<< B.readFile f