PGF.Haskell

Description

Auxiliary types and functions for use with grammars translated to Haskell with gf -output-format=haskell -haskell=concrete

Concrete syntax

class EnumAll a where

For enumerating parameter values used in tables

Methods

enumAll :: [a]

table :: (EnumAll k, Ord k) => [a] -> k -> a

Tables

type Str = [Tok]

Token sequences, output form linearization functions

data Tok

Tokens

Constructors

TK String

TP [([Prefix], Str)] Str

BIND

SOFT_BIND

CAPIT

type Prefix

Arguments

= String
To be matched with the prefix of a following token

fromStr :: Str -> String

Render a token sequence as a String

Common record types

class Has_s r a | r -> a where

Overloaded function to project the s field from any record type

Methods

proj_s :: r -> a

data R_s t

Haskell representation of the GF record type {s:t}

Constructors

R_s t

to_R_s :: Has_s r t => r -> R_s t

Coerce from any record type {...,s:t,...} field to the supertype {s:t}

Variants

(+++) :: Applicative f => f [a] -> f [a] -> f [a] infixr 5

Concatenation with variants

(!) :: Monad m => (t -> m (m a)) -> t -> m a

Selection from tables with variants

(!$) :: (Functor m, Monad m) => (a1 -> m a) -> m a1 -> m a

(!*) :: (Applicative m, Monad m) => m (a1 -> m a) -> m a1 -> m a