PGF.Lexing

Text lexing

lexText :: String -> [String]

Text lexing with standard word capitalization of the first word of every sentence

lexText' :: (String -> String) -> String -> [String]

Text lexing with custom treatment of the first word of every sentence.

unlexText :: [String] -> String

bindTok :: [String] -> [String]

Bind tokens separated by Prelude.BIND, i.e. &+

Code lexing

lexCode :: String -> [String]

Haskell lexer, usable for much code

unlexCode :: [String] -> String

lexLatexCode :: String -> [String]

LaTeX lexer in the math mode: should not be separated from the next word

Mixed lexing

lexMixed :: String -> [String]

LaTeX style lexer, with "math" environment using Code between $...$

unlexMixed :: [String] -> String

Additional lexing uitilties

capitInit :: [Char] -> [Char]

Capitalize first letter

uncapitInit :: [Char] -> [Char]

Uncapitalize first letter

unquote :: [[Char]] -> [[Char]]

Unquote each string wrapped in double quotes

isPunct :: Char -> Bool

isMajorPunct :: Char -> Bool

isMinorPunct :: Char -> Bool

isParen :: Char -> Bool

isClosing :: Char -> Bool