PGF2

Description

This is the Haskell binding to the C run-time system for loading and interpreting grammars compiled in Portable Grammar Format (PGF).

CId

type CId = String

PGF

data PGF

readPGF :: FilePath -> IO PGF

type AbsName

Arguments

= String
Name of abstract syntax

abstractName :: PGF -> AbsName

startCat :: PGF -> Cat

Concrete syntax

type ConcName

Arguments

= String
Name of concrete syntax

data Concr

languages :: PGF -> Map ConcName Concr

parse :: Concr -> Cat -> String -> Either String [(Expr, Float)]

parseWithHeuristics

Arguments

:: Concr
the language with which we parse

-> Cat
the start category

-> String
the input sentence

-> Double
the heuristic factor. A negative value tells the parser to lookup up the default from the grammar flags

-> [(Cat, Int -> String -> Int -> Maybe (Expr, Float, Int))]
a list of callbacks for literal categories. The arguments of the callback are: the index of the constituent for the literal category; the input sentence; the current offset in the sentence. If a literal has been recognized then the output should be Just (expr,probability,end_offset)

-> Either String [(Expr, Float)]

hasLinearization :: Concr -> Fun -> Bool

linearize :: Concr -> Expr -> String

alignWords :: Concr -> Expr -> [(String, [Int])]

Types

data Type

Constructors

DTyp [Hypo] CId [Expr]

type Hypo = (BindType, CId, Type)

Hypo represents a hypothesis in a type i.e. in the type A -> B, A is the hypothesis

functionType :: PGF -> CId -> Type

Trees

data Expr

type Fun

Arguments

= String
Name of function

readExpr :: String -> Maybe Expr

showExpr :: Expr -> String

mkApp :: Fun -> [Expr] -> Expr

unApp :: Expr -> Maybe (Fun, [Expr])

mkStr :: String -> Expr

Morphology

type MorphoAnalysis = (Fun, String, Float)

lookupMorpho :: Concr -> String -> [MorphoAnalysis]

fullFormLexicon :: Concr -> [(String, [MorphoAnalysis])]

Exceptions

newtype PGFError

Constructors

PGFError String

Grammar specific callbacks

type LiteralCallback = PGF -> (ConcName, Concr) -> Int -> String -> Int -> Maybe (Expr, Float, Int)

literalCallbacks :: [(AbsName, [(Cat, LiteralCallback)])]

Callbacks for the App grammar